requirements.txt
psycopg2-binary
scenedetect[opencv]
openai-whisper @ git+https://github.com/openai/whisper.git
ffmpeg-python
mediapipe
opencv-python
requests
